# algebra trick for much faster SEs in `get_predicted`