Today I was having trouble with `get_predicted`: it was taking a ton of time to produce estimates for a model with about 50K observations. So I decided to do a bit of profiling. It turns out that the vast majority of the time was spent on the “sandwich” multiplication.

In this PR, I introduce a simple matrix algebra trick for the sandwich computation, which yields roughly 1000x speedup in computation and 1000x reduction in memory load.

To begin, let’s create a variance-covariance matrix and a large model matrix by stacking `mtcars` multiple times:

``` r
library(bench)

mod <- lm(mpg ~ factor(cyl) + hp + drat, mtcars)
mtbig <- do.call("rbind", lapply(1:500, \(x) mtcars))
mm <- model.matrix(mod, data = mtbig)
vcovmat <- vcov(mod)
```

Before running the benchmark, let’s convince ourselves that the two strategies yield the same results at 10 digits tolerance:

``` r
old <- sqrt(diag(mm %*% vcovmat %*% t(mm)))
new <- sqrt(colSums(t(mm %*% vcovmat) * t(mm)))

identical(round(old, 10), round(new, 10))
# [1] TRUE
```

Run benchmark:

``` r
bench::mark(
    sqrt(diag(mm %*% vcovmat %*% t(mm))),
    sqrt(colSums(t(mm %*% vcovmat) * t(mm))))

#   expression                                    min   median `itr/sec` mem_alloc
#   <bch:expr>                               <bch:tm> <bch:tm>     <dbl> <bch:byt>
# 1 sqrt(diag(mm %*% vcovmat %*% t(mm)))        1.36s    1.36s     0.735    1.91GB
# 2 sqrt(colSums(t(mm %*% vcovmat) * t(mm))) 577.71µs   1.44ms   694.       1.95MB
```

After installing my branch, I tried `get_predicted` again on my problematic dataset, and the result came up almost instantaneously, whereas I had to wait a couple minutes before.