@bwiernik 2021-09-17
This looks great, thanks!

Could you check out the failing tests and see what the issue is? 

If possible, could you also write out the algebraic equivalence of the first line you changed as an equation? That would help me to double check