@vincentarelbundock 2021-09-17
I forced pushed a single line commit with the algebraic equivalence as a comment.

AFAICT, the error is the same that we have currently on `master`, so I don't think it's related to my commit. Looks like it's related to `brms`, but not 100% clear what the problem is.