@strengejacke 2021-09-17
_Should_ be fixed in https://github.com/easystats/insight/commit/c2bc2fbc99fc3eb17908984100b59da9a8ebd7e3, re-running checks now. Thanks for this commit, I used this in _ggeffects_ as well now :-)