@bwiernik 2021-09-17
Thanks! I was misreading the line you changed as M'VM instead of MVM'