@vincentarelbundock 2021-09-17
> Thanks for this commit, I used this in _ggeffects_ as well now :-)

The power of open source! ;-)